mtfc-generator/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
markdown>=3.4.0
typing-extensions>=4.8.0

